move custom artifact type import handling to standalone module
